+++ delphi-frontend/src/main/java/au/com/integradev/delphi/preprocessor/CompilerSwitchRegistry.java
@@ -22,14 +22,21 @@ import com.google.common.collect.Maps;
 import com.google.common.collect.Multimap;
 import com.google.common.collect.Multimaps;
 import com.google.common.collect.Range;
+import java.util.ArrayDeque;
+import java.util.Deque;
+import java.util.EnumSet;
 import java.util.HashSet;
+import java.util.Set;
+import javax.annotation.Nullable;
 import org.sonar.plugins.communitydelphi.api.directive.SwitchDirective.SwitchKind;
 
 public class CompilerSwitchRegistry {
   private final Multimap<SwitchKind, Range<Integer>> rangesBySwitchKind;
+  private final Deque<Set<SwitchKind>> savedSwitchStates;
 
   CompilerSwitchRegistry() {
     rangesBySwitchKind = Multimaps.newSetMultimap(Maps.newEnumMap(SwitchKind.class), HashSet::new);
+    savedSwitchStates = new ArrayDeque<>();
   }
 
   void addSwitch(SwitchKind kind, int startIndex, int endIndex) {
@@ -38,5 +45,16 @@ public class CompilerSwitchRegistry {
 
   public boolean isActiveSwitch(SwitchKind kind, int tokenIndex) {
     return rangesBySwitchKind.get(kind).stream().anyMatch(range -> range.contains(tokenIndex));
+  }
+
+  void pushState(Set<SwitchKind> activeSwitches) {
+    Set<SwitchKind> state = EnumSet.noneOf(SwitchKind.class);
+    state.addAll(activeSwitches);
+    savedSwitchStates.push(state);
+  }
+
+  @Nullable
+  Set<SwitchKind> popState() {
+    return savedSwitchStates.poll();
   }
 }